task2-feedback-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
jinja2==3.1.6
python-multipart==0.0.17
httpx==0.28.1
python-dotenv==1.2.1
psycopg2-binary==2.9.10
